requirements.txt
pandas
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
django-phone-field==1.8.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
EasyProcess==0.3
entrypoint2==0.2.3
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
geographiclib==1.50
geopy==2.1.0
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
install==1.3.4
ipykernel==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
macholib==1.14
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mpmath==1.1.0
mss==6.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.19.4
openpyxl==3.0.6
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
py==1.10.0
PyJWT==1.7.1
pynput==1.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.5
pyzmq==22.0.3
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
sympy==1.7.1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.3
webencodings==0.5.1
names
